친구 요청 / 수락 알람 기능 추가

See merge request s12-webmobile1-sub1/S12P11E203!82

+++ backend/aqoo/src/main/java/org/com/aqoo/domain/friend/dto/FriendResponse.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 
 @Data
 public class FriendResponse {
-    private Long id;            // 친구 관계 ID (FriendRelationship 테이블의 id)
+    private int id;            // 친구 관계 ID (FriendRelationship 테이블의 id)
     private String friendId;    // 친구의 ID (User 테이블의 id)
     private String nickname;    // 친구의 닉네임
     private String level;       // 친구의 레벨
     private String mainfish;    // 친구의 메인 피시
 
-    public FriendResponse(Long id, String friendId, String nickname, String level, String mainfish) {
+    public FriendResponse(int id, String friendId, String nickname, String level, String mainfish) {
         this.id = id;
         this.friendId = friendId;
         this.nickname = nickname;
